Fixed addFileToTree
Fixed the functionality of the method when adding multiple files to the tree that have the same data. Previously, no action was taken if the hash already existed in the tree, now no action is taken if the file is already in the tree. This allows for multiple files with the same data to exist in the tree.